Renamed  streaming_sequence_client to simple_sequence_client
